Refactor APIs

+++ SmartApparel/src/main/java/com/example/SmartApparel/Operations/dto/SalaryDTO.java
@@ -4,17 +4,41 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
+/**
+ * Data Transfer Object (DTO) class for Salary.
+ * This DTO class is used for transferring Salary data between different layers of the application.
+ */
 @NoArgsConstructor
 @AllArgsConstructor
 @Data
 public class SalaryDTO {
+
+    // Unique identifier for the salary record
     private int salaryId;
+
+    // Employee ID associated with the salary record
     private String empId;
+
+    // Status of the salary (e.g., paid, pending)
     private String status;
+
+    // Year and month for which the salary is applicable, in the format YYYY-MM
     private String yearNMonth;
-    private float basic;
-    private float epfByEmployee;
-    private float epfByCompany;
-    private float etfPayment;
-    private float netSalary;
+
+    // Basic salary amount
+    private BigDecimal basic;
+
+    // Employee's contribution to the Employees' Provident Fund (EPF)
+    private BigDecimal epfByEmployee;
+
+    // Company's contribution to the Employees' Provident Fund (EPF)
+    private BigDecimal epfByCompany;
+
+    // Employees' Trust Fund (ETF) payment
+    private BigDecimal etfPayment;
+
+    // Net salary after all deductions
+    private BigDecimal netSalary;
 }
